+++ src/main/java/com/yukari/relicera/common/item/BrutalPlunderBadgeItem.java
@@ -1,9 +1,6 @@
 package com.yukari.relicera.common.item;
 
-import net.minecraft.world.item.Item;
-import top.theillusivec4.curios.api.type.capability.ICurioItem;
-
-public class BrutalPlunderBadgeItem extends Item implements ICurioItem {
+public class BrutalPlunderBadgeItem extends QuickEquipCurioItem {
     public BrutalPlunderBadgeItem(Properties properties) {
         super(properties);
     }