+++ src/main/java/com/yukari/relicera/common/item/DivineSeveranceRingItem.java
@@ -1,8 +1,6 @@
 package com.yukari.relicera.common.item;
 
-import net.minecraft.world.item.Item;
-
-public class DivineSeveranceRingItem extends Item {
+public class DivineSeveranceRingItem extends QuickEquipCurioItem {
     public DivineSeveranceRingItem(Properties properties) {
         super(properties);
     }