+++ src/main/java/com/yukari/relicera/common/item/NightGlovesItem.java
@@ -1,9 +1,6 @@
 package com.yukari.relicera.common.item;
 
-import net.minecraft.world.item.Item;
-import top.theillusivec4.curios.api.type.capability.ICurioItem;
-
-public class NightGlovesItem extends Item implements ICurioItem {
+public class NightGlovesItem extends QuickEquipCurioItem {
     public NightGlovesItem(Properties properties) {
         super(properties);
     }